backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
demucs==4.0.1
pyrubberband==0.3.0
librosa==0.10.1
soundfile==0.12.1
google-generativeai==0.3.2
celery[redis]==5.3.6
redis==5.0.1
aiofiles==23.2.1
python-dotenv==1.0.0
python-magic-bin==0.4.14
